Update game for multiple level capability and dynamic maps
- Levels now have 5 working map parts
  - Platforms (the ground for the player)
  - Coins (the collectible objects). This may become multiple components
as I work in various prize points
  - Foreground: the non-interactive map that is behind the player
  - Background: the non-interactive map that is front of the player
  - Don't Touch: portions of the map that automatically restart the
player to start. This may see updates as enemies, health bars, and lives
come into gameplay
